Move Auth to Authorization namespace
