+++ Atamai.Slice.Sample/Slices/Session/Delete.cs
@@ -1,4 +1,4 @@
-using Atamai.Slice.Auth;
+using Atamai.Slice.Authorization;
 using Atamai.Slice.Swagger;
 
 namespace Atamai.Slice.Sample.Slices.Session;
