+++ Atamai.Slice/Authorization/AuthorizationMiddleware.cs
@@ -3,7 +3,7 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Atamai.Slice.Auth;
+namespace Atamai.Slice.Authorization;
 
 public static class AuthorizationMiddleware
 {
